refactor(boilerplate-vite): split src into client/ and server/, add sitemap
Restructure from src/ssr/ to explicit client/server split:
- src/client/entry.tsx — client hydration
- src/server/entry.tsx — SSR tree preparation
- src/server/server.ts — Express middleware with streaming
- src/server/sitemap.ts — sitemap getter for SitemapRenderer

Remove unused files:
- Shell.tsx (head injection done in server.ts, not via Shell component)
- renderer.tsx (empty, was kept for backward compat)

Add /sitemap.xml route using SitemapRenderer from @canonical/react-ssr.

Simplify build scripts: single build:server target.
