+++ apps/react/boilerplate-vite/index.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/ssr/entry-client.tsx"></script>
+    <script type="module" src="/src/client/entry.tsx"></script>
   </body>
 </html>
