+++ apps/react/boilerplate-vite/package.json
@@ -10,12 +10,8 @@
     "build:all": "bun run build:client && bun run build:server && bun run build:storybook",
     "build:storybook": "storybook build",
     "build:client": "vite build --ssrManifest --outDir dist/client",
-    "build:server": "bun run build:server:middleware",
-    "build:server:middleware": "vite build --ssr src/ssr/server.ts --outDir dist/server",
-    "build:server:cli": "vite build --ssr src/ssr/renderer.tsx --outDir dist/server",
-    "serve": "bun run serve:middleware",
-    "serve:middleware": "bun run build:client && bun run build:server:middleware && node dist/server/server.js",
-    "serve:cli": "bun run build:client && bun run build:server:cli && bun run serve-express dist/server/renderer.js",
+    "build:server": "vite build --ssr src/server/server.ts --outDir dist/server",
+    "serve": "bun run build:client && bun run build:server && node dist/server/server.js",
     "check": "bun run check:biome && bun run check:ts",
     "check:fix": "bun run check:biome:fix && bun run check:ts",
     "check:biome": "biome check",
