+++ apps/react/boilerplate-vite/src/client/entry.tsx
