+++ apps/react/boilerplate-vite/src/server/entry.tsx
