--- apps/react/boilerplate-vite/src/ssr/renderer.tsx
@@ -1,4 +1,0 @@
-// This file is kept for backward compatibility with the CLI render pipeline.
-// The primary SSR flow now uses renderToPipeableStream directly in server.ts.
-// See entry-server.tsx for the SSR tree setup.
-export {};
